requirements.txt
litellm>=1.40.0
llama-index-core>=0.10.45
llama-index-readers-file>=0.1.25
requests>=2.32.0
beautifulsoup4>=4.12.0
loguru>=0.6.0
pydantic>=2.7.0
PyYAML>=6.0
dirtyjson>=1.0.8
tiktoken>=0.7.0
